requirements.txt
allure-pytest>=2.8.6
pytest>=5.3.2
pytest-dependency>=0.4.0
python-dateutil>=2.8.2
selenium>=4.0.0
webdriver-manager>=3.5.2